requirements.txt
Django>=5.2
graphene-django>=3.2.3
django-graphql-jwt>=0.4.0 